Parse username and hostname from argv
This commit adds logic to parse arguments of the form `user@hostname`. Parsing
logic is currently very primitive and doesn't allow for the full set of
characters that may be present in a Windows username. A `winrm` package is also
added that contains a Host type that holds username, password, and hostname.