Use NTLM instead of HTTP Basic auth
This commit updates the WinRM connection used by the REPL to use NTLM instead
of HTTP Basic auth. This provides two benefits:

  - Windows account passwords aren't sent over the wire.

  - `winrm quickconfig` enables NTLM by default while Basic auth must be
    explicitly enabled.

Using NTLM does not provide any encryption for HTTP connections. This means
commands and responses are sent over the network in plaintext --- which
requires WinRM to be configured with AllowUnencrypted set to true. This is
unsafe for production use.